# Improve performance when generating with a specific locale