## Overview

This is a partial fix for the issue https://github.com/datafaker-net/datafaker/issues/1285.

Result before changes:
```
Time to process 1000000 values: 40.341209708s
```

Result after changes:
```
Time to process 1000000 values: 19.826039958s
```

Result with no locale:
```
Time to process 1000000 values: 2.310795584s
```

Results are based on the code here: https://github.com/vitaly-ivanov/datafaker-memory-leak/blob/main/app/src/main/kotlin/org/example/Performance.kt

## Details

Nulls were interpreted as no attempt to load

If values were missing for the specified locale, the `loadValues` method returned null instead of an empty map, resulting in subsequent attempts to load values into the cache.