@bodiam 2024-07-03
Thanks for this PR. Any chance you add a small unit / performance test to this to prevent regressions?