@asolntsev 2024-07-08
@bodiam @snuyanzin Could you **please release DF 2.3.1** with this fix?
I heard complains about performance of DS 2.3.0 (that became obvious after we fixed the memory leaks).